refactor: 간편 로그인 로직 수정
