+++ src/app/(auth)/signin/page.tsx
@@ -63,7 +63,7 @@ const SigninPage = () => {
   };
 
   const handleKakaoSigninClick = () => {
-    redirectKakaoAuth.signin();
+    redirectKakaoAuth();
   };
 
   const FORM_MARGIN_STYLES = 'mt-28 md:mt-42 lg:my-10';
