+++ src/lib/redirectKakaoAuth.ts
@@ -1,16 +1,5 @@
-export const redirectKakaoAuth = {
-  signin: () => {
-    redirect({ type: 'signin' });
-  },
-  signup: ({ state }: { state: string }) => {
-    redirect({ type: 'signup', state });
-  },
-};
-
-type AuthType = 'signin' | 'signup';
-
-const redirect = ({ type, state }: { type: AuthType; state?: string }) => {
-  const redirectUri = `${window.location.origin}/oauth/${type}/kakao/callback`;
+export const redirectKakaoAuth = () => {
+  const redirectUri = `${window.location.origin}/oauth/signup/kakao`;
   if (process.env.NODE_ENV === 'development') {
     const clientId = process.env.NEXT_PUBLIC_JS_KEY_KAKAO; // 환경변수에서
 
@@ -19,14 +8,12 @@ const redirect = ({ type, state }: { type: AuthType; state?: string }) => {
       `client_id=${clientId}&` +
       `redirect_uri=${encodeURIComponent(redirectUri)}&` +
       `response_type=code&` +
-      `${state ? `state=${encodeURIComponent(state)}&` : ''}` +
       `prompt=login`;
 
     window.location.href = kakaoAuthUrl;
   } else if (process.env.NODE_ENV === 'production') {
     Kakao.Auth.authorize({
       redirectUri,
-      ...(state && { state }),
     });
   }
 };
